# services/data-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "HTTP API: health/probe, кадры S3, инциденты и загруженность из ClickHouse, Swagger.",
        "title": "ITS Data Service API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/v1/cameras/{camera_id}/incidents": {
            "get": {
                "summary": "Инциденты по камере за интервал времени",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор камеры",
                        "name": "camera_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Начало интервала (RFC3339)",
                        "name": "from",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Конец интервала (RFC3339)",
                        "name": "to",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Лимит строк",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.RoadIncidentsListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/frames": {
            "get": {
                "summary": "Кадры камеры из S3 за интервал (ключи из ClickHouse + presigned URL)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор камеры",
                        "name": "camera_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Начало интервала (RFC3339)",
                        "name": "from",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Конец интервала (RFC3339)",
                        "name": "to",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Лимит строк",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.FramesListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/road_congestion": {
            "get": {
                "summary": "Список записей road_congestion",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Фильтр по segment_id",
                        "name": "segment_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Фильтр по camera_id",
                        "name": "camera_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Начало интервала (RFC3339)",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Конец интервала (RFC3339)",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Лимит строк",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.RoadCongestionListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/road_incidents": {
            "get": {
                "summary": "Список записей road_incidents",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Фильтр по segment_id",
                        "name": "segment_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Фильтр по camera_id",
                        "name": "camera_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Начало интервала (RFC3339)",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Конец интервала (RFC3339)",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Лимит строк (по умолчанию 100, макс. 500)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.RoadIncidentsListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/segments/{segment_id}/congestion/average": {
            "get": {
                "summary": "Средняя загруженность контура (segment) за интервал",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор контура/сегмента",
                        "name": "segment_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Начало интервала (RFC3339)",
                        "name": "from",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Конец интервала (RFC3339)",
                        "name": "to",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.CongestionAverageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "summary": "Liveness / health",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ProbeResponse"
                        }
                    }
                }
            }
        },
        "/probe/health": {
            "get": {
                "summary": "Health (probe path)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ProbeResponse"
                        }
                    }
                }
            }
        },
        "/probe/ready": {
            "get": {
                "summary": "Readiness probe",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ProbeResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.CongestionAverageResponse": {
            "type": "object",
            "properties": {
                "avg_congestion_score": {
                    "type": "number"
                },
                "from": {
                    "type": "string"
                },
                "sample_count": {
                    "type": "integer"
                },
                "segment_id": {
                    "type": "string"
                },
                "to": {
                    "type": "string"
                }
            }
        },
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "dto.FrameItem": {
            "type": "object",
            "properties": {
                "camera_id": {
                    "type": "string"
                },
                "download_url": {
                    "type": "string"
                },
                "observed_at": {
                    "type": "string"
                },
                "s3_key": {
                    "type": "string"
                },
                "segment_id": {
                    "type": "string"
                }
            }
        },
        "dto.FramesListResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.FrameItem"
                    }
                }
            }
        },
        "dto.ProbeResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
        "dto.RoadCongestionItem": {
            "type": "object",
            "properties": {
                "camera_id": {
                    "type": "string"
                },
                "congestion_score": {
                    "type": "number"
                },
                "observed_at": {
                    "type": "string"
                },
                "raw_ml": {
                    "type": "string"
                },
                "s3_key": {
                    "type": "string"
                },
                "segment_id": {
                    "type": "string"
                }
            }
        },
        "dto.RoadCongestionListResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.RoadCongestionItem"
                    }
                }
            }
        },
        "dto.RoadIncidentItem": {
            "type": "object",
            "properties": {
                "camera_id": {
                    "type": "string"
                },
                "crash_probability": {
                    "type": "number"
                },
                "incident_label": {
                    "type": "string"
                },
                "observed_at": {
                    "type": "string"
                },
                "raw_ml": {
                    "type": "string"
                },
                "s3_key": {
                    "type": "string"
                },
                "segment_id": {
                    "type": "string"
                }
            }
        },
        "dto.RoadIncidentsListResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.RoadIncidentItem"
                    }
                }
            }
        }
    }
}